deleted the user.js because it is a duplicate of account.js
