+++ models/Account.js
@@ -1,6 +1,5 @@
 const mongoose = require('mongoose');
-const bcrypt = require('bcrypt');
-const Transaction = require('../models/Transaction');
+
 // Define Account schema
 const AccountSchema = new mongoose.Schema(
     {
@@ -10,16 +9,23 @@ const AccountSchema = new mongoose.Schema(
           unique: true
          },
 
-       firstName: {
+       name: {
          type: String, 
          required: true
          },
-
-      lastName: {
-         type: String,
-          required: true
-         },
-
+         email:{
+          type: String,
+          required: true,
+          unique: true,
+          min: 6,
+          max: 255
+      },
+  password:
+   { type: String,
+       required: true,
+       min: 5,
+       max: 1024
+      },
       balance: { 
         type: Number, 
         required: true, 
@@ -40,7 +46,9 @@ const AccountSchema = new mongoose.Schema(
         type: mongoose.Schema.Types.ObjectId, 
         ref: 'Transaction' }]
   },
-  { timestamps: true }
+  { 
+    timestamps: true 
+  }
 );
 
 // export the model
