+++ routes/Account.js
@@ -8,7 +8,8 @@ const {
     getAccountTransactions,
     getAccountBalance,
     getAccountDeposit,
-    getAccountwithdrawals
+    getAccountwithdrawals,
+    Userlogin,
 } = require('../controllers/Account');
 
 const verifyToken = require ('../middleware/TokenHandler');
@@ -16,7 +17,7 @@ const verifyToken = require ('../middleware/TokenHandler');
 const router = express.Router();
 
 router.get("/", verifyToken, getAccounts);
-router.post("/", verifyToken, createAccount);
+router.post("/",  createAccount);
 router.get("/:id", verifyToken, getAccount);
 router.put("/:id", verifyToken, updateAccount);
 router.delete("/:id", verifyToken, deleteAccount);
@@ -24,6 +25,6 @@ router.get("/transactions/:id", verifyToken, getAccountTransactions);
 router.get("/balance/:id", verifyToken, getAccountBalance);
 router.post("/deposit/:id", verifyToken, getAccountDeposit);
 router.post("/withdraw/:id", verifyToken, getAccountwithdrawals);
-
+router.post("/login/", Userlogin);
 
 module.exports = router;
